Upload
The firmware relies on an external button to navigate through the states of the evaluation protocol.

Button is connected to Pin 2.

Pins 8, 12 and 13 are used to glow LEDs in accordance with the evaluation protocol state.

The ADC is fed by the EEG amplifier. The EEG amplifier's input is an EEG electrode that is placed on occipital part of skull (Oz position in 10-20 electrode system).